\name{absolute_freqs}
\alias{absolute_freqs}
\docType{data}
\title{
  A toy example dataset for mean measures of divergence, in a table format
}
\description{
  This artifical dataset includes 200 individuals described by 9 binary
  traits and splitted into 5 groups. To fit with commonly observed
  datasets in past sciences, a substantial amount of missing values have
  been added at random on this dataset.
}
\usage{data(absolute_freqs)}
\format{
  A matrix with 10 rows and 9 columns:
  \describe{
    \item{\code{Trait1}}{summary statistics for this trait}
    \item{\code{Trait2}}{summary statistics for this trait}
    \item{\code{Trait3}}{summary statistics for this trait}
    \item{\code{Trait4}}{summary statistics for this trait}
    \item{\code{Trait5}}{summary statistics for this trait}
    \item{\code{Trait6}}{summary statistics for this trait}
    \item{\code{Trait7}}{summary statistics for this trait}
    \item{\code{Trait8}}{summary statistics for this trait}
    \item{\code{Trait9}}{summary statistics for this trait}
  }
}
\keyword{datasets}
