\name{AnaSettings}
\alias{AnaSettings}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal function : tcltk code for a window which allow the user to specify some parameters for MMD analysis.
}
\description{
Internal function : tcltk code for a window which allow the user to specify some parameters for MMD analysis.
}
\usage{
AnaSettings(tab, type.data = "raw_data")
}
\arguments{
  \item{tab}{
The previously loaded dataframe.
}
  \item{type.data}{
"raw_data" or "summarized_data" : type of data selected by the user.
}
}
%- maybe also 'usage' for other objects documented here.
\author{
Frederic Santos
}
