\name{datafile_details}
\alias{datafile_details}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal function : tcltk code to display a window asking for details about the layout of the data file.
}
\description{
tcltk code for a windows asking if the data file contains raw binary data (standard case), or is a summarized matrix containing the number of individual and the proportion of presence for eacch variable and each group.
}
\usage{
datafile_details()
}
\value{
A character vector indicating the choice made by the user.
}
\author{
Frederic Santos
}
