\name{Protocole_MMD}
\alias{Protocole_MMD}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal function : tcltk code for a window which allow the user to specify some parameters for MMD analysis.
}
\description{
Internal function : tcltk code for a window which allow the user to specify some parameters for MMD analysis.
}
\usage{
Protocole_MMD()
}
%- maybe also 'usage' for other objects documented here.
\author{
Frederic Santos
}
