\name{Sel_type_data}
\alias{Sel_type_data}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal function : tcltk code for a window which allows the user to specify the type of the data.
}
\description{
Internal function : tcltk code for a window which allows the user to specify the type of the data.
}
\usage{
Sel_type_data()
}
\value{
"raw_data" or "summarized_data", depending on the radio button chosen by the user.
}
\author{
Frederic Santos
}

