\name{selvar}
\alias{selvar}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Internal function.
}
\description{
From a raw binary data set of n individuals belonging to one out of k groups, described by p variables ; constructs a summary mayrix giving the number of individuals and the proportion of "1" in each group and for each variable.
}
\usage{
selvar(T)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{T}{
A binary data frame or matrix, whose the first column is a group indicator.
}
}
\author{
Frederic Santos
}
