\name{toyMMD}
\alias{toyMMD}
\docType{data}
\title{
  A toy example dataset for mean measures of divergence, in a binary
  format
}
\description{
  This artifical dataset includes 200 individuals described by 9 binary
  traits and splitted into 5 groups. To fit with commonly observed
  datasets in past sciences, a substantial amount of missing values have
  been added at random on this dataset.
}
\usage{data(toyMMD)}
\format{
  A data frame with 200 observations on the following 10 variables:
  \describe{
    \item{\code{Group}}{a factor with 5 levels (group indicator)}
    \item{\code{Trait1}}{a numeric vector of zeroes and ones}
    \item{\code{Trait2}}{a numeric vector of zeroes and ones}
    \item{\code{Trait3}}{a numeric vector of zeroes and ones}
    \item{\code{Trait4}}{a numeric vector of zeroes and ones}
    \item{\code{Trait5}}{a numeric vector of zeroes and ones}
    \item{\code{Trait6}}{a numeric vector of zeroes and ones}
    \item{\code{Trait7}}{a numeric vector of zeroes and ones}
    \item{\code{Trait8}}{a numeric vector of zeroes and ones}
    \item{\code{Trait9}}{a numeric vector of zeroes and ones}
  }
}
\keyword{datasets}
