\name{AnthropMMD-package}
\alias{AnthropMMD-package}
\alias{AnthropMMD}
\docType{package}
\title{
A Shiny user interface for Smith's Mean Measure of Divergence.
}
\description{
Offers a Shiny user interface to work out Smith's Mean Measure of Divergence.
}
\details{
\tabular{ll}{
Package: \tab AnthropMMD\cr
Type: \tab Package\cr
Version: \tab 2.1.0\cr
Date: \tab 2017-08-09\cr
License: \tab GPL >=2\cr
}

}
\author{
Frederic Santos, <frederic.santos@u-bordeaux.fr>
}
\references{
Sjovold, T. (1977). Non-metrical Divergence Between Skeletal Populations. Ossa, 4, Supp. 1.

Harris, E. F., Sjovold, T. (2004). Calculation of Smith's Mean Measure of Divergence for Intergroup Comparisons Using Nonmetric Data. Dental Anthropology, 17(3), 83-93.
}
\keyword{MMD, Mean Measure of Divergence, Smith, nonmetric traits}
\examples{
\dontrun{ StartMMD() }
}
