\name{AnthropMMD-package}
\alias{AnthropMMD-package}
\alias{AnthropMMD}
\docType{package}
\title{
A GUI for Mean Measures of Divergence
}
\description{
Offers a complete and interactive GUI to work out a matrix of MMD from a binary dataset, especially for anthropologists.
}
\details{
\tabular{ll}{
Package: \tab AnthropMMD\cr
Type: \tab Package\cr
Version: \tab 0.9.7\cr
Date: \tab 2013-12-18\cr
License: \tab GPL (>= 2)\cr
}

For more information about this package, see the full (separated) documentation written by the author (contact by email).
}
\author{
Frederic Santos

Maintainer: Frederic Santos <frederic.santos@u-bordeaux1.fr>
}
\references{
Joel D. Irish. The mean measure of divergence : its utility in model-free and model-bound analyses relative to the mahalanobis d2 distance for nonmetric traits. American Journal of Human Biology, 2010.
}
\keyword{ MMD }

\examples{
\dontrun{StartMMD() # the only command to know : simply type it to begin !}
}
