\name{AnthropMMD-package}
\alias{AnthropMMD-package}
\alias{AnthropMMD}
\docType{package}
\title{
A GUI for Mean Measures of Divergence
}
\description{
Offers a complete and interactive GUI to work out a matrix of Mean Measures of Divergence from a binary dataset, especially for anthropologists.
}
\details{
\tabular{ll}{
Package: \tab AnthropMMD\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2016-04-07\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Frederic Santos

Maintainer: Frederic Santos <frederic.santos@u-bordeaux.fr>
}
\references{
Sjovold, T. (1977). Non-metrical Divergence Between Skeletal Populations. Ossa, 4, Supp. 1.

Harris, E. F., Sjovold, T. (2004). Calculation of Smith's Mean Measure of Divergence for Intergroup Comparisons Using Nonmetric Data. Dental Anthropology, 17(3), 83-93.
}
\keyword{ MMD }

\examples{
\dontrun{StartMMD() # the only command to know : simply type it to begin !}
}
