\name{AnthropMMD-package}
\alias{AnthropMMD-package}
\alias{AnthropMMD}
\docType{package}
\title{
A Shiny user interface for Smith's Mean Measure of Divergence (MMD)
}
\description{
Offers a Shiny user interface for the calculation of the Mean Measure of Divergence, with facilities for trait selection and graphical representations.
}
\details{
\tabular{ll}{
Package: \tab AnthropMMD\cr
Type: \tab Package\cr
Version: \tab 2.2.1\cr
Date: \tab 2017-09-11\cr
License: \tab GPL >=2\cr
}

}
\author{
Frederic Santos, <frederic.santos@u-bordeaux.fr>
}
\references{
Sjovold, T. (1977). Non-metrical Divergence Between Skeletal Populations. Ossa, \bold{4}, Supp. 1.

Harris, E. F. and Sjovold, T. (2004). Calculation of Smith's Mean Measure of Divergence for Intergroup Comparisons Using Nonmetric Data. Dental Anthropology, \bold{17}(3), 83--93.

Irish, J. (2010). The Mean Measure of Divergence: Its Utility in Model-Free and Model-Bound Analyses Relative to the Mahalanobis D2 Distance for Nonmetric Traits. American Journal of Human Biology, \bold{22}, 378--395.

Nikita, E. (2015). A Critical Review of the Mean Measure of Divergence and Mahalanobis Distances Using Artificial Data and New Approaches to the Estimation of Biodistances Employing Nonmetric Traits. American Journal of Physical Anthropology, \bold{157}, 284--294.
}
\keyword{MMD, Mean Measure of Divergence, Smith, nonmetric traits}
\examples{
\dontrun{ StartMMD() }
}
